Introduce ADRIO system of Geo creation
The Active Data Resource Interface Object (ADRIO) system streamlines the process of Geo creation by allowing for the creation of .geo files that specify the data needed by a Geo. The data are automatically fetched and cached when a simulation is run with one of these files. Data can also be pre-cached using the new 'cache' CLI command.