Collect simulation data in two forms: by home and by visit. (#208)
We have been tabulating simulation data according to the place where any individual happens to be during the simulation tick. That is, "by visit".
It can also be useful to tabulate this data by the individuals' home location -- that is, "by home". We now do both.
An Output object continues to default to "by visit" data, but you can flip this mode by calling `out.data_by_home` which returns another Output object that contains the same data but now it defaults to the "by home" data.
In usage, you will select your data mode this way first, then chain a call to `plot.line()` or another such method.
Regardless of which mode the Ouptut object is in, it is still possible to explicitly access either data set however. `events_by_visit`, for example.