`sim_messaging()` "liveness" improvements.
If not overridden by setting `live=True/False` when creating the context, it now tries to auto-detect environments in which static rendering is more appropriate. Thus, users have less reason to override it.

In addition, there is the EPYMORPH_LIVE_MESSAGING env var to allow users to override the default behavior for the environment without code changes.

This also adds the settings module with some foundational utilities for implementing epymorph configuration.