Fix issue with geo scope labels and the LabeledLocations initializer. (#269)
Geo scope labels were not always available during parameter resolution. This lead to failures using the LabeledLocations initializer.

Both RUME resolution and SimFunc resolution now provide a last-line fallback for any "label" attribute using the geo scope labels (if available). The Database system was simplified in the process.

The existing "meta::geo::label" system was found to not do anything, so it was removed to avoid confusion.

* Clean USAGE notebook exec.

* Fix doc. Hide as_rume_params method.